refactor: split MCP server into state, helpers, tools, and entry point modules
server.py was 472 lines with all 10 tools, helpers, and state in one
file. Now split into:
- state.py: GraphState, load_graph, auto-reload, require_graph
- helpers.py: compact_entity, compact_relationship
- tools.py: All 10 @mcp.tool() functions via register_tools()
- server.py: Slim entry point (~30 lines)

Closes #32
